Omit pepr install for _helpers

+++ _helpers/dev/cli.mts
@@ -76,10 +76,12 @@ const test = program.command('test')
     process.env.PEPR_IMAGE = thisCommand.opts().image
 
     try {
-      execSync(`npm i ${getPeprAlias()}`)
-      execSync('rm package-lock.json', {cwd: peprExcellentExamplesRepo})
+      if(path.basename(process.cwd()) !== '_helpers'){
+        execSync(`npm i ${getPeprAlias()}`)
+        execSync('rm package-lock.json', {cwd: peprExcellentExamplesRepo})
+      }
+
       execSync('npm install', { cwd: peprExcellentExamplesRepo });
-      console.log(execSync('grep "pepr" package-lock.json', {cwd: peprExcellentExamplesRepo}).toString())
     } catch (err) {
       throw new Error(`Failed to run npm install in ${peprExcellentExamplesRepo}: ${err.message}`);
     }
